Refactor Product and ProductRepo to move the stock from a product to ProductRepo Refactor and fix modifyOrder to work as intended

+++ src/main/java/com/phaete/ShopService.java
@@ -21,13 +21,16 @@ public class ShopService {
      * Prints a message to the console if the product does not exist.
      *
      * @param order The order to be checked.
+     * @return true if all the products in the order exist, false if not.
      */
-    public void productOfOrderExists(Order order) {
+    public boolean productOfOrderExists(Order order) {
         for (Product product : order.products().keySet()) {
             if (productRepo.getProduct(product.id()) == null) {
                 System.out.println("Product does not exist: " + product.id());
+                return false;
             }
         }
+        return true;
     }
 
     /**
@@ -35,13 +38,16 @@ public class ShopService {
      * Prints a message to the console for each product that does not have enough stock.
      *
      * @param order The order to be checked.
+     * @return true if all the products in the order have enough stock, false if not.
      */
-    public void productStockOfOrderExists(Order order) {
+    public boolean productStockOfOrderExists(Order order) {
         for (Product product : order.products().keySet()) {
-            if (!productRepo.getProduct(product.id()).hasEnoughStock(order.products().get(product))) {
+            if (!productRepo.hasEnoughStock(product, order.products().get(product))) {
                 System.out.println("Product is not in stock: " + product.id());
+                return false;
             }
         }
+        return true;
     }
 
     /**
@@ -51,12 +57,14 @@ public class ShopService {
      * @param order The order to be placed
      */
     public void placeOrder(Order order) {
-        orderRepo.addOrder(order);
-        for (Product product : order.products().keySet()) {
-            productRepo.decreaseStock(
-                    productRepo.getProduct(product.id()),
-                    order.products().get(product)
-            );
+        if (productOfOrderExists(order) && productStockOfOrderExists(order)) {
+            orderRepo.addOrder(order);
+            for (Product product : order.products().keySet()) {
+                productRepo.decreaseStock(
+                        productRepo.getProduct(product.id()),
+                        order.products().get(product)
+                );
+            }
         }
     }
 
@@ -66,18 +74,36 @@ public class ShopService {
      *
      * @param order The order to be removed
      */
-    public void removeOrder(Order order) {
-        orderRepo.removeOrder(order);
-        for (Product product : order.products().keySet()) {
-            productRepo.increaseStock(
-                    productRepo.getProduct(product.id()),
-                    order.products().get(product)
-            );
+    public void cancelOrder(Order order) {
+        if (productOfOrderExists(order) && productStockOfOrderExists(order)) {
+            orderRepo.removeOrder(order);
+            for (Product product : order.products().keySet()) {
+                productRepo.increaseStock(
+                        productRepo.getProduct(product.id()),
+                        order.products().get(product)
+                );
+            }
         }
     }
 
     public boolean modifyOrder(Order order, Map<Product, Integer> modifications) {
-        return orderRepo.modifyOrder(order, modifications);
+        // Check if all products in the order can be changed before actually changing anything
+        for (Product product : modifications.keySet()) {
+            if (!productRepo.hasEnoughStock(product, modifications.get(product))) {
+                System.out.println("Product is not in high enough stock available to apply the modification: " + product.id() + ". Left over stock: " + productRepo.getProductRepo().get(product));
+                return false;
+            }
+        }
+        // Cancel the old order
+        cancelOrder(order);
+        // Create a new order and change it based on the modifications
+        Order modifiedOrder = new Order(order.id(), order.products(), order.customerName(), order.orderDate());
+        for (Product product : modifiedOrder.products().keySet()) {
+            modifiedOrder.addProduct(product, modifications.get(product));
+        }
+        // Place the new order, which is the old order with the modifications applied
+        placeOrder(modifiedOrder);
+        return true;
     }
 
     public OrderRepo getOrderRepo() {
